fix: int16 parsing bug
Appears to have lived here for 5 years. Not fully clear how it actually worked before.

+++ Dwarf/Read.cs
@@ -20,7 +20,7 @@ namespace ElfParser.Dwarf
                     break;
                 case DW_FORM.Block2:
                     {
-                        var numBytes = BitConverter.ToInt32(infoData.GetRange(index, 2).ToArray(), 0);
+                        var numBytes = BitConverter.ToInt16(infoData.GetRange(index, 2).ToArray(), 0);
                         index += 2;
                         output.AddRange(infoData.GetRange(index, numBytes));
                         index += numBytes;
